refactor poem selector into separate component
