+++ src/components/InputText.jsx
@@ -8,13 +8,11 @@ import FormControlLabel from '@mui/material/FormControlLabel';
 import FormControl from '@mui/material/FormControl';
 import FormLabel from '@mui/material/FormLabel';
 import InputLabel from '@mui/material/InputLabel';
-import MenuItem from '@mui/material/MenuItem';
 import Radio from '@mui/material/Radio';
 import RadioGroup from '@mui/material/RadioGroup';
-import Select from '@mui/material/Select';
 import useMediaQuery from '@mui/material/useMediaQuery';
 
-import AuthorProgress from './AuthorProgress'
+import PoemSelector from './PoemSelector'
 import SnackbarAlerts from './SnackbarAlerts'
 
 import "./InputText.css";
@@ -23,7 +21,7 @@ import {PARSERS as P,
         tagWordsInLine,
         ParserDescriptions} from './ParserDescriptions'
 
-import sonnets, { 
+import sonnets, {
     defaultAuthor, defaultAuthorList,
     defaultTitle, defaultTitleList,
     defaultTextLines,
@@ -48,10 +46,6 @@ const UNICODE_NBSP = "\u00A0"
 
 const PRE = "word" // prefix for identifying word spans by computed id
 
-function capitalizeFirstLetter(string) {
-    return string.charAt(0).toUpperCase() + string.slice(1);
-}
-
 function InputText(props) {
 
     // my suspicion is that when I added a sufficiently large number of
@@ -68,9 +62,7 @@ function InputText(props) {
     const [authorList, setAuthorList] = useState(defaultAuthorList)
     const [title, setTitle] = useState(defaultTitle)
     const [titleList, setTitleList] = useState(defaultTitleList)
-
     const [toast, setToast] = useState({open: false, severity: "info", message: ""})
-
     const [loadingProgress, setLoadingProgress] = useState({author: "", percentage: 0})
 
     const extraLargeScreen = useMediaQuery(theme => theme.breakpoints.up('xl'));
@@ -79,11 +71,11 @@ function InputText(props) {
     const [nounInverters, setNounInverters] = useState({
         [P.PARTS_OF_SPEECH]: [],
         [P.EN_POS]: [],
-    }) 
+    })
     // nounInverter values are jagged arrays of arrays indicating if a word at a
     //    given index (in a given line)
     //    is an inversion from the noun-state by the parserName
-    //    e.g. nounInverters[P.PARTS_OF_SPEECH][3][7] === true 
+    //    e.g. nounInverters[P.PARTS_OF_SPEECH][3][7] === true
     //      means that the eighth word (7+1) in the fourth (3+1) line (for the
     //      P.PARTS_OF_SPEECH parser) has been inverted
     //    i.e. inverted means now should be considered not-a-noun if originally
@@ -216,32 +208,6 @@ function InputText(props) {
         setParserName(value)
     }
 
-    // generic selector used for authors and titles
-    function selector(selName, value, setter, valList) {
-        return (
-            <FormControl dense="true">
-              <InputLabel id={`${selName}-select-label`}>{capitalizeFirstLetter(selName)}</InputLabel>
-              <Select
-                labelId={`${selName}-select-label`}
-                id={`${selName}-select`}
-                value={value}
-                label={capitalizeFirstLetter(selName)}
-                onChange={ (event) => setter(event.target.value) }
-              >
-                {valList.map( s => (<MenuItem value={s}>{s}</MenuItem>) )}
-              </Select>
-            </FormControl>
-        )
-    }
-
-    function authorSelector() {
-        return selector('author', author, setAuthor, authorList)
-    }
-
-    function titleSelector() {
-        return selector('title', title, setTitle, titleList)
-    }
-
     function toastAlert(message, severity) {
         setToast({message, severity, open: true})
     }
@@ -286,7 +252,7 @@ function InputText(props) {
 
             // Choose the 2nd poet just because we want it to be Emily Dickinson in our common case
             // or 1st (0th) if we end up with a one poet list
-            const authorIndex = Math.min(1, authors.length-1) 
+            const authorIndex = Math.min(1, authors.length-1)
 
             // Set titles first because it is reverse order of triggering updates to selectors
             const titles = titlesByAuthor[authors[authorIndex]]
@@ -310,7 +276,7 @@ function InputText(props) {
         if (titlesByAuthor !== undefined) {
             const titles = titlesByAuthor[author]
             setTitleList(titles)
-            setTitle(titles[0])
+            setTitle(titles ? titles[0] : "")
         }
     }, [author] )
 
@@ -318,61 +284,61 @@ function InputText(props) {
         // When the title changes, the entire poem changes
         if (fetchedPoems !== undefined) {
             // reset the noun inverter so this it will be reconstructed for the current parser
-            updateValueForCurrentParser(setNounInverters, []) 
+            updateValueForCurrentParser(setNounInverters, [])
             setTextLines(fetchedPoems[author][title])
         }
     }, [title] )
 
-  return (
-    <section>
-        <Grid container spacing={2} direction={extraLargeScreen?"row":"column"}>
-            <Grid item xs={6}>
-              <h1> Select a poem </h1>
-              { authorSelector() }
-              { titleSelector() }
-              { loadingProgress.percentage > 0 &&
-                loadingProgress.percentage < 100 && 
-                <AuthorProgress {...loadingProgress}/>  }
-              <textarea value={textLines && textLines.join("\n")} id="text-input"/>
-            </Grid>
-            <Grid item xs={6}>
-                <h1> Choose your Natural Language Parser </h1>
-                <div>
-                    <ParserDescriptions />
-                    <FormControl>
-                        <FormLabel id="parsers-radio-buttons-group-label">Parsers</FormLabel>
-                        <RadioGroup
-                            row
-                            aria-labelledby="parsers-radio-buttons-group-label"
-                            defaultValue={defaultParserName}
-                            name="parserName"
-                            value={parserName}
-                            onChange={handleParserChange}
-                        >
-                            <FormControlLabel value="parts-of-speech" control={<Radio />} label="Parts-of-Speech" />
-                            <FormControlLabel value="en-pos" control={<Radio />} label="en-pos" />
-                        </RadioGroup>
-                    </FormControl> 
-                </div>
-                <h1> Correct what is and is not a noun </h1>
-                <ul>
-                    <li>Click on a word with the <span className="non-noun">Plus</span> <img id="non-noun-cursor-img"></img> cursor to change a word INTO a noun.</li>
-                    <li>Click on a word with the <span className="noun">Back</span> <img id="noun-cursor-img"></img> cursor to change a word BACK TO a non-noun.</li>
-                </ul>
-                <div id="text-output"></div>
-                <fieldset id="stats-fieldset">
-                    <legend id="stats-legend"><b><i>Statistics for {parserName}</i></b></legend>
-                    <div>
-                        <span><b>False Positives:</b> {falsePositiveCount[parserName]}  </span>
-                        <span><b>False Negatives:</b> {falseNegativeCount[parserName]}</span> 
-                    </div>
-                    <b>Total Incorrect:</b> {falsePositiveCount[parserName] + falseNegativeCount[parserName]}
-                </fieldset>
-            </Grid>
-        </Grid>
-      <SnackbarAlerts {...{...toast, setSnackOpen}}/>
-    </section>
-  )
+    const poemSelectionCriteria = {
+        author, setAuthor, authorList,
+        title, setTitle, titleList,
+        loadingProgress, textLines
     }
+
+    return (
+        <section>
+          <Grid container spacing={2} direction={extraLargeScreen?"row":"column"}>
+            <Grid item xs={6}>
+              <PoemSelector {...poemSelectionCriteria} />
+            </Grid>
+            <Grid item xs={6}>
+              <h1> Choose your Natural Language Parser </h1>
+              <div>
+                <ParserDescriptions />
+                <FormControl>
+                  <FormLabel id="parsers-radio-buttons-group-label">Parsers</FormLabel>
+                  <RadioGroup
+                    row
+                    aria-labelledby="parsers-radio-buttons-group-label"
+                    defaultValue={defaultParserName}
+                    name="parserName"
+                    value={parserName}
+                    onChange={handleParserChange}
+                  >
+                    <FormControlLabel value="parts-of-speech" control={<Radio />} label="Parts-of-Speech" />
+                    <FormControlLabel value="en-pos" control={<Radio />} label="en-pos" />
+                  </RadioGroup>
+                </FormControl>
+              </div>
+              <h1> Correct what is and is not a noun </h1>
+              <ul>
+                <li>Click on a word with the <span className="non-noun">Plus</span> <img id="non-noun-cursor-img"></img> cursor to change a word INTO a noun.</li>
+                <li>Click on a word with the <span className="noun">Back</span> <img id="noun-cursor-img"></img> cursor to change a word BACK TO a non-noun.</li>
+              </ul>
+              <div id="text-output"></div>
+              <fieldset id="stats-fieldset">
+                <legend id="stats-legend"><b><i>Statistics for {parserName}</i></b></legend>
+                <div>
+                  <span><b>False Positives:</b> {falsePositiveCount[parserName]}  </span>
+                  <span><b>False Negatives:</b> {falseNegativeCount[parserName]}</span> 
+                </div>
+                <b>Total Incorrect:</b> {falsePositiveCount[parserName] + falseNegativeCount[parserName]}
+              </fieldset>
+            </Grid>
+          </Grid>
+          <SnackbarAlerts {...{...toast, setSnackOpen}}/>
+        </section>
+    )
+}
 
 export default InputText;
